+++ tests/src/Functional/EmailTest.php
@@ -46,9 +46,12 @@ class EmailTest extends ApigeeEdgeFunctionalTestBase {
    */
   protected function tearDown() {
     try {
-      $this->developer->delete();
+      if ($this->developer !== NULL) {
+        $this->developer->delete();
+      }
     }
     catch (\Exception $exception) {
+      $this->logException($exception);
     }
     parent::tearDown();
   }
